requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.1.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
joblib>=1.3.0
gunicorn>=21.0.0
pytest>=7.4.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
